backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pydantic-settings
python-dotenv
requests
beautifulsoup4
google-api-python-client
google-auth-oauthlib
google-auth-httplib2
celery
redis
pytest
httpx
rank-bm25==0.2.2
tiktoken==0.7.0
openai==1.51.0
alembic==1.13.3
tqdm==4.66.5
pgvector==0.3.3
